pr fixes rename green to event
