+++ components/buttons/stories/index.js
@@ -17,7 +17,7 @@ import StoryButtonsPrevNext from './StoryButtonsPrevNext.vue';
 import StoryButtonCTA from './StoryButtonCTA.vue';
 import StoryButtonAdjacent from './StoryButtonAdjacent.vue';
 import StoryButtonAdjacentFullWidth from './StoryButtonAdjacentFullWidth.vue';
-import StoryButtonGreen from './StoryButtonGreen.vue';
+import StoryButtonEvent from './StoryButtonEvent.vue';
 
 import ButtonDefault from './ButtonDefault.vue';
 /* ##Import story component here */
@@ -32,7 +32,7 @@ storiesOf('Buttons', module)
   .add('Inverted button', createStory(StoryButtonInverted))
   .add('Prev/Next aligned buttons', createStory(StoryButtonsPrevNext))
   .add('CTA variant', createStory(StoryButtonCTA))
-  .add('Green', createStory(StoryButtonGreen));
+  .add('Event', createStory(StoryButtonEvent));
 
 storiesOf('Buttons/Adjacent', module)
   .add('Default', createStory(StoryButtonAdjacent))
